+++ src/game/units/wizard/spells/wall_of_stone/mod.rs
@@ -2,3 +2,4 @@ pub(crate) mod components;
 pub(crate) mod constants;
 pub(super) mod plugin;
 pub(crate) mod systems;
+pub(crate) mod wall_material;